+++ foundation/application.go
@@ -2,7 +2,6 @@ package foundation
 
 import (
 	"context"
-	"encoding/json"
 	"flag"
 	"os"
 	"path/filepath"
@@ -206,39 +205,6 @@ func (app *Application) registerCommands(commands []consolecontract.Command) {
 
 func (app *Application) setTimezone() {
 	carbon.SetTimezone(app.MakeConfig().GetString("app.timezone", carbon.UTC))
-}
-
-type JSON struct {
-	marshal   func(any) ([]byte, error)
-	unmarshal func([]byte, any) error
-}
-
-func NewJSON() supportcontract.Json {
-	if Json == nil {
-		Json = &JSON{
-			marshal:   json.Marshal,
-			unmarshal: json.Unmarshal,
-		}
-	}
-
-	return Json
-}
-
-func (j *JSON) Marshal(v any) ([]byte, error) {
-	return j.marshal(v)
-}
-
-func (j *JSON) MarshalString(v any) (string, error) {
-	s, err := j.marshal(v)
-	return string(s), err
-}
-
-func (j *JSON) Unmarshal(data []byte, v any) error {
-	return j.unmarshal(data, v)
-}
-
-func (j *JSON) UnmarshalString(data string, v any) error {
-	return j.unmarshal([]byte(data), v)
 }
 
 func setEnv() {
